% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rpart.lists}
\alias{rpart.lists}
\title{Creates lists of variable values (factor levels) associated with each rule in an \code{\link{rpart}} object.}
\usage{
rpart.lists(object)
}
\arguments{
\item{object}{an rpart object}
}
\value{
a list of lists
}
\description{
Creates lists of variable values (factor levels) associated with each rule in an \code{\link{rpart}} object.
}
\examples{
library(rpart)
fit<-rpart(Reliability~.,data=car90)
rpart.lists(fit)
}

