% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rpart.rules.push}
\alias{rpart.rules.push}
\title{Writes rule tables required to process rpart rules in SQL to an open RODBC connection.}
\usage{
rpart.rules.push(object, connection, prefix = "")
}
\arguments{
\item{object}{an rpart object}

\item{connection}{and open RODBC connection}

\item{prefix}{A character string to prefix the rules to allow for multiple rule sets}
}
\description{
Writes rule tables required to process rpart rules in SQL to an open RODBC connection.
}

