% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rpart.rules}
\alias{rpart.rules}
\title{Returns a list of strings summarizing the branch path to each node.}
\usage{
rpart.rules(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
Returns a list of strings summarizing the branch path to each node.
}
\examples{
library(rpart)
fit<-rpart(Reliability~.,data=car90)
rpart.rules(fit)
}

