% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rpart.rules.table}
\alias{rpart.rules.table}
\title{Returns an unpivoted table of branch paths (subrules) associated with each node.}
\usage{
rpart.rules.table(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
Returns an unpivoted table of branch paths (subrules) associated with each node.
}
\examples{
library(rpart)
fit <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)
rpart.rules.table(fit)
}

