% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rpart.subrules.table}
\alias{rpart.subrules.table}
\title{Returns an unpivoted table of variable values (factor levels) associated with each branch.}
\usage{
rpart.subrules.table(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
Returns an unpivoted table of variable values (factor levels) associated with each branch.
}
\examples{
library(rpart)
fit <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)
rpart.subrules.table(fit)
}

