% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rpart.lists}
\alias{rpart.lists}
\title{Creates lists of variable values (factor levels) associated with each rule in an rpart object.}
\usage{
rpart.lists(object)
}
\arguments{
\item{object}{an rpart object}
}
\value{
a list of lists
}
\description{
Creates lists of variable values (factor levels) associated with each rule in an rpart object.
}
\examples{
library(rpart)
fit <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)
rpart.lists(fit)
}

